+++ src/main/java/firok/tiths/block/punji/BlockPunjiNetherrack.java
@@ -1,0 +1,5 @@
+package firok.tiths.block.punji;
+
+public class BlockPunjiNetherrack extends BlockPunjiCustom
+{
+}