+++ src/main/java/firok/tiths/block/punji/BlockPunjiObsidian.java
@@ -1,0 +1,5 @@
+package firok.tiths.block.punji;
+
+public class BlockPunjiObsidian extends BlockPunjiCustom
+{
+}