+++ src/main/java/firok/tiths/common/Items.java
@@ -37,7 +37,7 @@ public final class Items
 	@Reg(nameAncientBookIrisia)
 	public static final Item ancientBookIrisia = new ItemXP(true, 20); // 艾瑞西亚古籍
 	@Reg(nameBeetrootStack)
-	public static final Item beetrootStack = new ItemCustom(); // 甜菜捆
+	public static final Item beetrootStack = new ItemBait(Blocks.blockBaitBeetrood); // 甜菜捆
 	@Reg(value = nameBrumeJade, od = {"gem", "jade"})
 	public static final Item brumeJade = new ItemCustom(); // 云玉
 	@Reg(prefBelt+nameTraitHeavy)
@@ -71,7 +71,7 @@ public final class Items
 	@Reg("buoyant_light")
 	public static final Item buoyantLight = new ItemBuoyantLight(); // 浮灯
 	@Reg(nameCarrotStack)
-	public static final Item carrotStack = new ItemCustom(); // 萝卜捆
+	public static final Item carrotStack = new ItemBait(Blocks.blockBaitCarrot); // 萝卜捆
 	@Reg("charm_soul_watcher")
 	public static final Item charmSoulWatcher = new ItemSoulWatchBauble(BaubleType.CHARM).setRarity(EnumRarity.UNCOMMON);
 	@Reg(nameChloroplastDressing)
@@ -116,7 +116,7 @@ public final class Items
 	public static final Item fleshCooked = new ItemFood(14, 1f, false)
 			.setPotionEffect(new PotionEffect(MobEffects.STRENGTH, 600, 0), 1); // 熟血肉
 	@Reg(nameFishStack)
-	public static final Item fishStack = new ItemCustom(); // 鱼捆
+	public static final Item fishStack = new ItemBait(Blocks.blockBaitFish); // 鱼捆
 	@Reg(nameFluidBall)
 	public static final ItemFluidBall fluidBall = new ItemFluidBall(); // 液体球
 	@Reg(nameFluidCollector)
@@ -282,7 +282,7 @@ public final class Items
 	public static final Item ringSoulGather = new ItemSoulGatherBauble(BaubleType.RING, 1, 0);
 	public static final Item ringSoulGatherCreative = new ItemSoulGatherBauble(BaubleType.RING, 100, 0).setRarity(TithsRarity.CreativeOnly);
 	@Reg(nameSeedStack)
-	public static final Item seedStack = new ItemCustom(); // 种子捆
+	public static final Item seedStack = new ItemBait(Blocks.blockBaitSeed); // 种子捆
 	@Reg(nameShell)
 	public static final Item shell = new ItemCustomFood(1, 0.2f, false); // 贝壳
 	@Reg(nameShellCooked)
@@ -317,7 +317,7 @@ public final class Items
 	@Reg(nameVibratingCrystal)
 	public static final Item vibratingCrystal = new ItemCustom(); // 振晶
 	@Reg(nameWheatStack)
-	public static final Item wheatStack = new ItemCustom(); // 小麦捆
+	public static final Item wheatStack = new ItemBait(Blocks.blockBaitWheat); // 小麦捆
 	@Reg(nameWitheringEssence)
 	public static final Item witheringEssence = new ItemCustom().setRarity(EnumRarity.UNCOMMON); // 凋零精华