+++ changelog.md
@@ -2,6 +2,15 @@
 
 > All Time Stamps: GMT+8
 
+## 0.3.17.0 _2020-08-13 23:30_
+
+* 增加方块 added blocks
+  * 铁钉 iron punji block
+  * 海绵钉 sponge punji block
+  * 云钉 cloud punji block
+* 增加状态效果 added potion effects
+  * 坚冰 hard ice
+
 ## 0.3.16.0 _2020-08-10 20:00_
 
 * 移除方块 removed blocks